Fix conditional compilation
